refactor: move V4 exceptions into common exceptions
